+++ src/main/java/com/musseukpeople/woorimap/tag/application/dto/request/TagRequest.java
@@ -1,4 +1,4 @@
-package com.musseukpeople.woorimap.tag.application.dto;
+package com.musseukpeople.woorimap.tag.application.dto.request;
 
 import javax.validation.constraints.NotBlank;
